requirements.txt
Flask==1.1.2
uuid==1.30
boto3==1.18.65
psycopg2-binary==2.9.1
openai==0.10.2
re==2.2.1
subprocess32==3.5.4
dash==2.0.0
numpy==1.21.2
pandas==1.3.3
waitress==2.0.0
Flask-Session==0.3.2
authlib==0.15.2
python-dotenv==0.19.1
